First pass at Helm support (#12)
Very first pass at supporting Helm. If, when crawling directories, we find a Helm chart, we use the Helm library to render it and pass the rendered YAML files to the regular loader.

Also, refactor the loader so that there's exactly one LintContext per directory, instead of having one in total. The issue with having just one LintContext is that checks which depend on other objects become more expensive, and we risk making mistakes. (For example, if you're checking whether a service has a label that matches a pod, you may accidentally think it does match a pod if it matches a pod in some other random Helm chart/directory, even if it doesn't match any pod that's actually deployed with it.)