Fix test that checks MaxSurge on DaemonSet with MaxUnavailable
A former case `daemonSetWithStrategy` was likely added by mistake.
DaemonSet was configured with `MaxUnavailable` but the test
`TestMinSurgeInteger` worked by validating `MaxSurge`.
Still the test passed due to former behavior in a dependency that
did not report `MaxSurge` setting to be present for DaemonSet.
Note that if it would have been a Deployment instead of DaemonSet,
the test would fail.
After updating dependencies, the behavior of DaemonSet's MaxSurge
became consistent with Deployment and so the test case started
failing.
Per clarification with Rob Rati in CoreOS slack DM, the failure is
valid, therefore I modify this test expectation.

Example of failing test
https://app.circleci.com/pipelines/github/stackrox/kube-linter/688/workflows/06748c44-bfe6-4254-b593-1c98edcaf834/jobs/2798

Also, I added couple more cases around DaemonSet and Deployment in
the same test for completeness. The similar ones should ideally be
added to other tests but that's too much work :-) and too much noise
to already lenghty tests.