Add tests to verify object paths for loaded helm charts
Archive and renamed directory tests fail at this point. The latter is
noted in the issue https://github.com/stackrox/kube-linter/issues/212

Note that when testing with modified chart directory name, I
temporarily rename the directory and then rename it back at the end
of the test.
I do this because I don't see a simple way to copy directory tree in
Go without writing own implementation. :-(
Hopefully test will not crash often leaving renamed directory, or
developer will notice and not commit files if it happens to crash.