Create new check for liveness port
Want to ensure the port is correctly set for liveness probe since it's
easy to make a typo when people are doing copy pasta. This will let you
know it's failing airier and not have to wait for a failed deployment.

Want to do this also for startup and readiness checks. Assume it make
sense to have them as other checks and was planning on doing that in a
follow up.